refactor: RankManager -> RankResult 이름 변경 및 Rank 클래스 vo -> domain 변경
