+++ src/test/java/lotto/domain/rank/RankTest.java
@@ -1,5 +1,6 @@
-package lotto.domain.rank.vo;
+package lotto.domain.rank;
 
+import lotto.domain.rank.Rank;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
